refactor(cli): extract CLI parse/validation helpers + tests from main.rs
main.rs was 2196 lines. Move the argument parse/validation helpers
(`parse_mic_silence_ms`, `TtsChoice` + `From`, `validate_speech_assets`,
`warn_on_npu_serialisation`, `npu_serialisation_warning`,
`pair_reasoning_markers`) into a new `cli_support` module, and the four
`#[cfg(test)]` modules into `cli_support/tests.rs`. main.rs drops to 1357
lines (-839) and now holds only the runtime wiring + the `Cli` struct.

The `Cli` struct stays in the crate root with all field visibilities
unchanged: the moved helpers reach its private fields through
descendant-module visibility (a child module can read an ancestor's
private items), so no `pub(crate)` field churn was needed.

Also fix a pre-existing latent non-exhaustive match: `validate_speech_assets`
only covered Piper/Supertonic/MacosNative, but `TtsBackend::AndroidNative`
(added in the recent android-native work) made the portable
`--features speech` build fail to compile. The two unreachable
platform-native arms are now collapsed into `MacosNative | AndroidNative`.
Zero behaviour change — the CLI's `TtsChoice` can only produce
Piper/Supertonic.

No behaviour change otherwise; relocation is verbatim. Verified across
default / speech / speech+macos-native / qnn (clippy + tests) and the full
workspace test suite, all green; fmt clean.
